tests/wpt/tests/css/css-lists/parsing/counter-increment-valid.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>CSS Lists: parsing counter-increment with valid values</title>
<link rel="help" href="https://drafts.csswg.org/css-lists-3/#propdef-counter-increment">
<meta name="assert" content="counter-increment supports the full grammar '[ <counter-name> <integer>? ]+ | none'.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/parsing-testcommon.js"></script>
</head>
<body>
<script>
test_valid_value('counter-increment', 'none');
test_valid_value('counter-increment', 'chapter', 'chapter 1');
test_valid_value('counter-increment', 'section -1');
test_valid_value('counter-increment', 'first -1 second third 99', 'first -1 second 1 third 99');
</script>
</body>
</html>
